Developer

+++ social-network/backend/java/src/main/java/com/ourproject/socialnetwork/service/UserService.java
@@ -1,40 +1,50 @@
 package com.ourproject.socialnetwork.service;
 
-import com.ourproject.socialnetwork.entity.User;
+import com.ourproject.socialnetwork.mapper.UserMapper;
+import com.ourproject.socialnetwork.model.UserDto;
 import com.ourproject.socialnetwork.repository.UserRepository;
+import jakarta.persistence.EntityNotFoundException;
 import java.util.List;
-import java.util.NoSuchElementException;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.dao.DuplicateKeyException;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 @Service
 public class UserService {
     private final UserRepository userRepository;
 
     @Autowired
-    public UserService(UserRepository userRepository) {
+    public UserService(UserRepository userRepository, SequenceGeneratorService userSequenceService) {
         this.userRepository = userRepository;
     }
 
-    public List<User> findAllUser() {
-        return userRepository.findAll();
+    public List<UserDto> findAllUser() {
+        var users = userRepository.findAll();
+        return users.stream()
+                .map(UserMapper.INSTANCE::userToUserDto)
+                .toList();
     }
 
-    public User getUserByUserName(String name) {
-        var user = userRepository.findUserByUsername(name);
-        return user.orElseThrow(() -> new NoSuchElementException("User not found with name " + name));
+    public UserDto getUserByUserName(String name) throws EntityNotFoundException {
+        var optionalUser = userRepository.findUserByUsername(name);
+        var user = optionalUser.orElseThrow(() -> new EntityNotFoundException("User not found with name " + name));
+        return UserMapper.INSTANCE.userToUserDto(user);
     }
 
-    public User addUser(User user) throws DuplicateKeyException {
-        return userRepository.insert(user);
+    @Transactional
+    public UserDto updateUser(UserDto userUpdate) throws EntityNotFoundException {
+        var user = UserMapper.INSTANCE.userDtoToUser(userUpdate);
+        if (!userRepository.existsById(user.getUserId())) {
+            throw new EntityNotFoundException("User not found with id: " + user.getUserId());
+        }
+        user = userRepository.save(user);
+        return UserMapper.INSTANCE.userToUserDto(user);
     }
 
-    public User updateUser(User user) throws DuplicateKeyException {
-        return userRepository.save(user);
-    }
-
-    public void deleteUser(Long id) throws DuplicateKeyException {
+    public void deleteUser(Long id) throws EntityNotFoundException {
+        if (!userRepository.existsById(id)) {
+            throw new EntityNotFoundException("User with id: " + id + " doesn't exist");
+        }
         userRepository.deleteById(id);
     }
 }
